+++ src/classBased/sezzleDOMFunctions.js
@@ -1,52 +1,15 @@
 /* eslint-disable class-methods-use-this */
-import RenderAwesomeSezzle from './renderAwesomeSezzle';
+// import RenderAwesomeSezzle from './renderAwesomeSezzle';
 import Utils from './utils';
-import Modal from './modal';
 
 class sezzleDOMFunctions {
-  constructor() {
-    this._configInst = null;
-    this._modalInst = new Modal();
-    this._renderAwesomeSezzleInst = new RenderAwesomeSezzle();
-    this._allConfigsUsePriceClassElement = true;
-    this._els = [];
-    this._intervalInMs = 2000;
-    // Functions required in RenderAwesomeSezzle, will be sent as param! Didn't make sense to put in utils
-    this.funtionsForRenderSezzle = {
-      getPriceText: this.getPriceText,
-      isProductEligible: this.isProductEligible,
-      getFormattedPrice: this.getFormattedPrice,
-      getElementsByXPath: this.getElementsByXPath,
-    };
+  constructor(config) {
+    this._configInst = config;
   }
 
   /**
    * ************* PUBLIC FUNCTIONS ***************
   */
-
-  /**
-   * All steps required to show the widget
-  */
-  init(_configInst) {
-    this._configInst = _configInst;
-
-    this._configInst.configGroups.forEach((configGroup, index) => {
-      if (configGroup.hasPriceClassElement) {
-        this._renderAwesomeSezzleInst.render(
-          configGroup.priceElements[0], configGroup.renderElements[0],
-          0, index,
-          this.funtionsForRenderSezzle, this._configInst,
-        );
-        this._startObserve(configGroup.priceElements[0], (mutations) => {
-          this._mutationCallBack(mutations, index);
-        });
-      } else {
-        this._allConfigsUsePriceClassElement = false;
-      }
-    });
-    if (!this._allConfigsUsePriceClassElement) this._sezzleWidgetCheckInterval();
-    this._modalInst.renderModals(this._configInst);
-  }
 
   /**
    * This function fetches all the elements that is pointed to by the given xpath
@@ -113,6 +76,63 @@ class sezzleDOMFunctions {
 
     children = children.filter((c) => c !== null);
     return this.getElementsByXPath(xpath, xindex + 1, children);
+  }
+
+  /**
+   * This function finds out the element where Sezzle's widget
+   * will be rendered. By default it would return the parent element
+   * of the given price element. If the over ride path is found and
+   * it leads to a valid element then that element will be returned
+   * @param element - This is the price element
+   * @param index - Index of the config group that element belongs to
+   * @return the element where Sezzle's widget will be rendered
+  */
+  getElementToRender(element, index = 0) {
+    let toRenderElement = null;
+    if (this._configInst.configGroups[index].rendertopath !== null) {
+      const path = Utils.breakXPath(this._configInst.configGroups[index].rendertopath);
+      toRenderElement = element;
+      for (let i = 0; i < path.length; i++) {
+        const p = path[i];
+        if (toRenderElement === null) {
+          break;
+        } else if (p === '.') {
+          // eslint-disable-next-line no-continue
+          continue;
+        } else if (p === '..') {
+          // One level back
+          toRenderElement = toRenderElement.parentElement;
+        } else if (p[0] === '.') {
+          // The class in the element
+          toRenderElement = toRenderElement.getElementsByClassName(p.substr(1)).length
+            ? toRenderElement.getElementsByClassName(p.substr(1))[0]
+            : null;
+        } else if (p[0] === '#') {
+          // The ID in the element
+          toRenderElement = document.getElementById(p.substr(1));
+        } else if (p === '::first-child') {
+          // rendered as first child
+          toRenderElement = toRenderElement.children.length > 0
+            ? toRenderElement.firstElementChild
+            : null;
+          this._configInst.configGroups[index].widgetIsFirstChild = true;
+        } else {
+          // If this is a tag
+          // indexes are 0-indexed (e.g. span-2 means third span)
+          let indexToTake = 0;
+          if (p.split('-').length > 1) {
+            if (p.split('-')[1] >= 0) {
+              indexToTake = parseInt(p.split('-')[1], 10);
+            }
+          }
+          toRenderElement = toRenderElement.getElementsByTagName(p.split('-')[0]).length > indexToTake
+            ? toRenderElement.getElementsByTagName(p.split('-')[0])[indexToTake]
+            : null;
+        }
+      }
+    }
+    // return the element's parent if toRenderElement is null
+    return toRenderElement || element.parentElement;
   }
 
   /**
@@ -266,126 +286,6 @@ class sezzleDOMFunctions {
         }
       });
     }
-  }
-
-  /**
-   * Looks for newly added price elements
-  */
-  _sezzleWidgetCheckInterval() {
-    this._configInst.configGroups.forEach((configGroup, index) => {
-      if (configGroup.xpath === []) return;
-      const elements = this.getElementsByXPath(configGroup.xpath);
-      elements.forEach((e) => {
-        if (!e.hasAttribute('data-sezzleindex')) {
-          this._els.push({
-            element: e,
-            toRenderElement: this._getElementToRender(e, index),
-            deleted: false,
-            observer: null,
-            configGroupIndex: index,
-          });
-        }
-      });
-    });
-    // add the sezzle widget to the price elements
-    this._els.forEach((el, index) => {
-      if (!el.element.hasAttribute('data-sezzleindex')) {
-        const sz = this._renderAwesomeSezzleInst.render(
-          el.element, el.toRenderElement,
-          index, el.configGroupIndex, this, this._configInst,
-        );
-        if (sz) {
-          el.observer = this._startObserve(el.element, (mutations) => {
-            this._mutationCallBack(mutations, el.configGroupIndex);
-          });
-          this._modalInst.addClickEventForModal(sz, el.configGroupIndex);
-          this._observeRelatedElements(el.element, sz, this._configInst.configGroups[el.configGroupIndex].relatedElementActions);
-        } else {
-          // remove the element from the els array
-          delete this._els[index];
-        }
-      }
-    });
-    // refresh the array
-    this._els = this._els.filter((e) => e !== undefined);
-    // Find the deleted price elements
-    // remove corresponding Sezzle widgets if exists
-    this._els.forEach((el, index) => {
-      if (el.element.parentElement === null && !el.deleted) { // element is deleted
-        // Stop observing for changes in the element
-        if (el.observer !== null) el.observer.disconnect();
-        // Mark that element as deleted
-        el.deleted = true;
-        // Delete the corresponding sezzle widget if exist
-        const tmp = document.getElementsByClassName(`sezzlewidgetindex-${index}`);
-        if (tmp.length) {
-          const sw = tmp[0];
-          sw.parentElement.removeChild(sw);
-        }
-      }
-    });
-    // Hide elements ex: afterpay
-    for (let index = 0, len = this._configInst.configGroups.length; index < len; index++) {
-      this._hideSezzleHideElements(index);
-    }
-    setTimeout(() => this._sezzleWidgetCheckInterval(), this._intervalInMs);
-  }
-
-  /**
-   * This function finds out the element where Sezzle's widget
-   * will be rendered. By default it would return the parent element
-   * of the given price element. If the over ride path is found and
-   * it leads to a valid element then that element will be returned
-   * @param element - This is the price element
-   * @param index - Index of the config group that element belongs to
-   * @return the element where Sezzle's widget will be rendered
-  */
-  _getElementToRender(element, index = 0) {
-    let toRenderElement = null;
-    if (this._configInst.configGroups[index].rendertopath !== null) {
-      const path = Utils.breakXPath(this._configInst.configGroups[index].rendertopath);
-      toRenderElement = element;
-      for (let i = 0; i < path.length; i++) {
-        const p = path[i];
-        if (toRenderElement === null) {
-          break;
-        } else if (p === '.') {
-          // eslint-disable-next-line no-continue
-          continue;
-        } else if (p === '..') {
-          // One level back
-          toRenderElement = toRenderElement.parentElement;
-        } else if (p[0] === '.') {
-          // The class in the element
-          toRenderElement = toRenderElement.getElementsByClassName(p.substr(1)).length
-            ? toRenderElement.getElementsByClassName(p.substr(1))[0]
-            : null;
-        } else if (p[0] === '#') {
-          // The ID in the element
-          toRenderElement = document.getElementById(p.substr(1));
-        } else if (p === '::first-child') {
-          // rendered as first child
-          toRenderElement = toRenderElement.children.length > 0
-            ? toRenderElement.firstElementChild
-            : null;
-          this._configInst.configGroups[index].widgetIsFirstChild = true;
-        } else {
-          // If this is a tag
-          // indexes are 0-indexed (e.g. span-2 means third span)
-          let indexToTake = 0;
-          if (p.split('-').length > 1) {
-            if (p.split('-')[1] >= 0) {
-              indexToTake = parseInt(p.split('-')[1], 10);
-            }
-          }
-          toRenderElement = toRenderElement.getElementsByTagName(p.split('-')[0]).length > indexToTake
-            ? toRenderElement.getElementsByTagName(p.split('-')[0])[indexToTake]
-            : null;
-        }
-      }
-    }
-    // return the element's parent if toRenderElement is null
-    return toRenderElement || element.parentElement;
   }
 
   /**
